feat(manager, manager_admin): move system deletion endpoint from manager to manager-admin
VULN-1943
