+++ manager.admin.spec.yaml
@@ -413,6 +413,26 @@ paths:
                           description: Percentage of free space
                           example: 49.12
 
+  /database/systems/{inventory_id}:
+    delete:
+      summary: Delete system.
+      description: Deletes system from database.
+      operationId: manager.admin_handler.DeleteSystem.delete
+      x-methodName: deleteSystem
+      security:
+        - ApiKeyAuthAdmin: []
+      responses:
+        200:
+          description: System has been deleted.
+        404:
+          description: System has not been found in the database.
+          content:
+            application/vnd.api+json:
+              schema:
+                $ref: '#/components/schemas/Errors'
+      parameters:
+        - $ref: '#/components/parameters/inventory_id'
+
 
 components:
   parameters:
