+++ manager.spec.yaml
@@ -536,31 +536,6 @@ paths:
                 $ref: '#/components/schemas/Errors'
       parameters:
         - $ref: '#/components/parameters/inventory_id'
-    delete:
-      summary: Delete system.
-      description: Deletes system from database.
-      operationId: manager.system_handler.DeleteSystem.delete
-      x-methodName: deleteSystem
-      security:
-        - ApiKeyAuth: []
-        - BasicAuth: []
-      responses:
-        200:
-          description: System has been deleted.
-        404:
-          description: System has not been found in the database.
-          content:
-            application/vnd.api+json:
-              schema:
-                $ref: '#/components/schemas/Errors'
-        503:
-          description: Service is running in read-only mode.
-          content:
-            application/vnd.api+json:
-              schema:
-                $ref: '#/components/schemas/Errors'
-      parameters:
-        - $ref: '#/components/parameters/inventory_id'
 
   /systems/{inventory_id}/cves:
     get:
