+++ manager/admin_handler.py
@@ -8,7 +8,7 @@ from peewee import DatabaseError
 
 from common.config import Config
 from common.logging import get_logger
-from common.peewee_model import ContentVersion, DB
+from common.peewee_model import ContentVersion, DB, SystemPlatform
 from .base import DeleteRequest, GetRequest, PutRequest, ApplicationException
 
 CFG = Config()
@@ -300,3 +300,24 @@ class VacuumUser(GetRequest):
             LOGGER.exception("Internal server error: %s", err)
             return "Error"
         return "Ok"
+
+
+class DeleteSystem(DeleteRequest):
+    """DELETE to /v1/database/systems/{inventory_id}"""
+
+    _endpoint_name = r"/v1/database/systems/{inventory_id}"
+
+    @classmethod
+    def _system_exists(cls, inventory_id):
+        system = (SystemPlatform.select()
+                  .where((SystemPlatform.inventory_id == inventory_id)))
+        if not system.exists():
+            raise ApplicationException("inventory_id must exist", 404)
+
+    @classmethod
+    def handle_delete(cls, **kwargs):
+        """Delete system"""
+        inventory_id = kwargs["inventory_id"]
+        cls._system_exists(inventory_id)
+        DB.execute_sql("SELECT delete_system(%s)", (inventory_id,))
+        return ""
