+++ manager/system_handler.py
@@ -10,10 +10,10 @@ from peewee import Case, DoesNotExist, fn, JOIN, SQL
 from common.config import Config
 from common.logging import get_logger
 from common.peewee_model import BusinessRisk, CveAccountData, CveMetadata, SystemPlatform, \
-    SystemVulnerabilities, CveImpact, Status, DB, RHAccount, InsightsRule, InventoryHosts
+    SystemVulnerabilities, CveImpact, Status, RHAccount, InsightsRule, InventoryHosts
 from common.utils import external_service_request
 from .base import ApplicationException, DEFAULT_BUSINESS_RISK, GetRequest, OS_INFO_QUERY, OS_INFO_SORT, parse_int_list, PatchRequest, \
-    DeleteRequest, CVE_SYNOPSIS_SORT, parse_str_list, parse_str_or_list, reporter, unique_bool_list, parse_tags, \
+    CVE_SYNOPSIS_SORT, parse_str_list, parse_str_or_list, reporter, unique_bool_list, parse_tags, \
     cyndi_join, ServiceAccess, get_advisories_per_cve_from_db, filter_by_advisory, \
     get_account_data, has_advisory, remediation
 from .filters import apply_filters, filter_types
@@ -565,26 +565,3 @@ class PatchBulkSystemsOptOut(PatchRequest):
         if not updated:
             raise ApplicationException('inventory_id must exist and inventory_id must be visible to user', 404)
         return {'updated': updated}
-
-
-class DeleteSystem(DeleteRequest):
-    """DELETE to /systems/{inventory_id}"""
-
-    _endpoint_name = r'/v1/systems/{inventory_id}'
-
-    @classmethod
-    def _system_exists_and_visible(cls, inventory_id):
-        system = (SystemPlatform.select()
-                  .join(RHAccount, on=(SystemPlatform.rh_account_id == RHAccount.id))
-                  .where((RHAccount.name == connexion.context['user']) &
-                         (SystemPlatform.inventory_id == inventory_id)))
-        if not system.exists():
-            raise ApplicationException("inventory_id must exist and inventory_id must be visible to user", 404)
-
-    @classmethod
-    def handle_delete(cls, **kwargs):
-        """Delete system"""
-        inventory_id = kwargs['inventory_id']
-        cls._system_exists_and_visible(inventory_id)
-        DB.execute_sql("SELECT delete_system(%s)", (inventory_id,))
-        return ""
