+++ tests/manager_tests/test_admin_handler.py
@@ -185,3 +185,14 @@ class TestAdminManager(FlaskTestCaseAdmin):
         monkeypatch.setattr(DB, "execute_sql", lambda _, __: MockDbRes())
         monkeypatch.setattr(subprocess, "Popen", self._vacuum_table_user_mock)
         self.vfetch("database/vacuum/user/1")
+
+    def test_system_delete(self):
+        """Test deletion of valid system"""
+        self.vfetch("/database/systems/00000000-0000-0000-0000-000000000004", data="", method="DELETE")
+        response = self.vfetch("/database/systems/00000000-0000-0000-0000-000000000004", data="", method="DELETE")
+        assert response.status_code == 404
+
+    def test_system_delete_invalid_id(self):
+        """Test deletion of invalid system"""
+        response = self.vfetch("/database/systems/b1f21450-0000-1111-2222-000000000000", data="", method="DELETE").check_response(status_code=404)
+        assert response.status_code == 404
