+++ tests/manager_tests/test_readonly.py
@@ -22,8 +22,3 @@ class TestReadOnly(FlaskTestCase):
         monkeypatch.setattr(CFG, 'read_only_mode', 1)
         response = self.vfetch("systems/00000000-0000-0000-0000-000000000004/opt_out?value=true", data="", method="PATCH")
         assert response.status_code == 503
-
-    def test_delete(self, monkeypatch):
-        monkeypatch.setattr(CFG, 'read_only_mode', 1)
-        response = self.vfetch("systems/00000000-0000-0000-0000-000000000004", method="DELETE")
-        assert response.status_code == 503
