+++ tests/manager_tests/test_system_handler.py
@@ -202,15 +202,6 @@ class TestSystemHandler(FlaskTestCase):
         response = self.vfetch('systems/b1f21450-0000-1111-2222-000000000000').check_response(status_code=404)
         assert response.body.errors[0].detail == 'inventory_id must exist and inventory_id must be visible to user'
 
-    def test_system_delete(self):
-        self.vfetch("systems/00000000-0000-0000-0000-000000000004", data="", method="DELETE").check_response()
-        response = self.vfetch("systems/00000000-0000-0000-0000-000000000004", data="", method="DELETE").check_response(status_code=404)
-        assert response.body.errors[0].detail == 'inventory_id must exist and inventory_id must be visible to user'
-
-    def test_system_delete_invalid_id(self):
-        response = self.vfetch("systems/b1f21450-0000-1111-2222-000000000000", data="", method="DELETE").check_response(status_code=404)
-        assert response.body.errors[0].detail == 'inventory_id must exist and inventory_id must be visible to user'
-
     def test_system_business_risk_filtering(self):
         response = self.vfetch('systems/00000000-0000-0000-0000-000000000005/cves?business_risk_id=0').check_response()
         assert len(response.body.data) == 7
